Azure Pipelines: install Qt 5.12.7 (LTS) and include AppImage updater
Currently when we run "Help -> Check for Updates" in a continuous AppImage, we get: "You are running a self-compiled Git version on a platform without official snapshots, and we're unable to determine if a new version is out.".

This commit updates Qt to 5.12.7 (LTS) and includes the AppImageUpdaterBridge Qt Plugin by @antony-jr.

It features binary delta updates, which means that going from continuous build to continuous build the user does not have to re-download the full application, but only the bits that have actually changed.

Note that thanks to the work of @antony-jr all of this is working as a Qt plugin, without even touching any Mumble source code.